* Resolved several compiler warnings on Windows
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@167 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ core/src/queue_kevent.c
@@ -107,7 +107,7 @@ _dispatch_mgr_invoke(dispatch_queue_t dq)
 			FD_COPY(&_dispatch_rfds, &tmp_rfds);
 			FD_COPY(&_dispatch_wfds, &tmp_wfds);
 			if (timeoutp) {
-				sel_timeout.tv_sec = timeoutp->tv_sec;
+				sel_timeout.tv_sec = (long)timeoutp->tv_sec;
 #if defined(__GNUC__)
 				sel_timeout.tv_usec = (typeof(sel_timeout.tv_usec))(timeoutp->tv_nsec / 1000u);
 #else
